requirements.txt
requests>=2.32.3
beautifulsoup4==4.12.2
lxml==4.9.3
pinecone==7.0.2
openai>=1.0.0
playwright>=1.40.0
python-dotenv>=1.0.0
flask>=2.3.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0 